requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
biopython==1.81
numpy==1.24.3
pydantic==2.5.0
requests==2.32.5
